WIP add u8ident
Extension for secure utf-8 identifiers, for everybody
seeing identifiable names in UTF-8 encodings, like
a filename in a terminal or UI widget.
Identifiers need to be identifiable, i.e. implement mixed
script detection, and such for a Unicode TR39
Moderately Restrictive restriction level.
Also identifiers are validated and normalized by default, to
be able to compare and find them.