uset: change iterator WIP TEMP
need two states: if next is not empty, we are still in the bucket chain.
if empty, we need to advance to the next bucket.

storing the hash and advancing it by one is wrong. it will stail fail
on rehashing during iteration.
storing a global index is too slow, as we have to re-iterate on every step.

add lot of temp lot of uset logging and asserts. needs to reverted eventually