analyzing perf_lst_iterate, add -s for silent
perf record/perf report:
don't see any bad code at all.
the only reason why it could be slower, is that
that stack variables for sum and it are farther away in C than in C++.
offset 0x50 and 0x60, vs in 0 and rcx reg in C++.

so restore -O3 by default.