iter redesign step 3: deque
embed the value and additional container/index in the returned iter/T*.
for now only for deque, which is the hardest.

we don't need foreach_ref then, as the value is just *ref/*iter, and
(I, next) is trivial.
add (I, advance).
This looks now very close to the C++ STL.
Added optimized foreach variants for the header, and generic variants
for algorithm into iterators.

The deque iterator is now faster than the C++ STL one (!), even if
we do two assignments, updating ref and the index. So no next links
neede between the buckets.