fixup set 2way compare
support both way and 3way compare with set.find and set.insert.
this was only caught by generate!
do formal verification for both. this would have caught it also.

fix set.generate to replace all.
nodes with equal values might move.

performance: insert/erase are now a bit slower than the STL (before we were faster),
but now we are safer. also much safer iterators, and they are still faster than the STL.
set.clear is now as slow as in the STL