add obscure deque generate_n tests, and insert_range.
According to our definition of generate_n, without an output
iterator and a front_inserter. One can argue which is more generate_n.
STL inserts via an inserter, and enlarges the container.
CTL foreach n skips the output iterator overflowing the size.

deque.insert_range is broken for the STL when the range is empty. skip that.