likewise for INCLUDE_ALGORITHM
a bit bothersome to declare,
but helps with the compile benchmark.

One exeception is string.h, as this only loaded once.
and if the first of those didn't request INCLUDE_ALGORITHM
we would need to reload it later, which is guarded by
ifndef __CTL_STRING__H__ and loaded from within vector.
pretty complicated, which we avoid for now. See the end of string.h for test code.